Move to the first page when the feedback page size is changed
When the user changes the page size in the feedback list
(filter feedback form results), move to the first page in the list.

If the user were currently on the last page and
the page number were kept, then increasing the page size would
cause the 404 not found error because
the same page number would not exist any longer.
There are less pages in total when each page contains more
list items.

Fixes #37