Update to use Django 4.2
Remove features deprecated and removed in Django versions 4.0, 4.1 and
4.2.
Also update dependencies to use versions that support Django 4.

Note: In the file `dynamic_forms/forms.py` there still appears a call of
the function `_html_output,` which is deprecated as of Django 4.0.
This is marked with a TODO comment. This should be fixed before or when
updating to Django 5.

Fix #49