Display static texts also after student has submitted
Change LabelField to be "disabled", indicating it's uneditable.
LabelField, which is used to display static text between questions
is not editable, therefore can always use the initial value.
Before the static text was dropped out after the student has submitted,
this fixes the issue and always includes the static text in the form.
Fixes #20