Add context tags
Create a new model "ContextTag", which the user can use to create
tags, which are displayed to represent answers to determined questions.
For example, this can be used by to generate tags to indicate which
option a student responded to a the question of how well they understood
the material.
This is part of the improvement of how the feedback responses are
displayed.