Fix rendering of older responses to students
If a teacher responded to or edited an response to an older version
of feedback, the response would not appear for the student in the
material unless the teacher would respond afterwards to the the newest
feedback or if a student sent a new response.

Change the uploading so if the feedback being responded to is an older
version, Aplus also updates the newes version, which is displayed by
the exercise.