Add script for updating new, empty database fields after v2.6
In the version upgrade v2.6, new fields were added to the database
tables. Some tables must be updated manually afterwards so that
existing records are updated with the correct values for
the new attributes.

This commit adds a script that fills in the values for
`exercise.parent_name`. The correct `parent_name` is retrieved
from the A+ REST API.

Fixes #75