Fix colortag buttons behavior
The feedback tag colortag buttons (in the feedback list view) used
to be spans, although they perform as buttons. This is an accessibility
issue, but also had the side effect of those not being focusable
(problematic in keyboard navigation).
Change the element type from span to button.

Also apply clickhandler 'ajax_set_tag_state' to only colortags that
are buttons, not all colortags (so remove click handler and the
resulting warning from from colortags that shouldn't be clickable.

Fixes #63