Update django-colortag to fix tag filter urls
Previously, when filtering feedback based on student and feedback tags,
the url was constructed using ordinal numbers for the tags. Change this use tag slugs.
This means that the url won't change for the same query if new tags are added
(which previously occured).
The change was implemented my Mikael Lenander in the django-colortag
repository.

Fixes #91