Display relevant points in a popover by feedback
Implement popover, which displays points that are relevant for the
context of the given feedback.
Adds a new view (with URL) that calculates the points for the student
for the entire course (by categories), as well as the module and the
chapter that the feedback was given from. This content is inserted into
the popover.
The popover displays the points in both numerical format as well as
progress bars for the points by category, module and chapter.
The total points are only displayed numerically.

Requires A+ points API to provide max_points and
max_points_by_difficulty for the course.

Fixes #65