Show text responses immediately above text fields for students
Change student view so the responses sent by the student are shown
immediately above the text field and empty the text field.
This should make it clearer to the student that sending another response
does not edit the previous response but sends a new message.
This also means that messages don't need to indicate which field the
response was from (using the field key).

If there are several text fields, display teacher responses by the
"primary" text feedback field, which is either a text field with the
flag "main-feedback" (if any field has the flag) or the last text field.

Change also the modal submission preview so it highlights the last
message from the student with a blue border.

Add aria-label to icon indicating response is from teacher.
Add screen-reader-only caption to communicate that the "well" contains
previous responses and teacher responses.

Fixes #54
Fixes #52