Place grade-and-send buttons in segmented dropdown button
Change the radio button (+ submit) replacement buttons from a block
button group to a segmented button with dropdown toggle.
Have the "Good" (full points) option be the default option visible
and the two other options available in the dropdown.

This makes it more compact, so all the other contents in the response
message bottom bar should fit on the same row, and there is room also
for styling buttons.