Fix total conversation count when paginated
Previously the total conversation count still was only the number of
conversations on the displayed page. Change the total_conversation_count
to actually have all of the filtered conversations (not just the ones
on the page.)

Fixup commit to cf39a47712e74fb9611ab84ae60ef53bfed6ef9a.
Fixes #94