Change submitter heading layout
Change submitter heading (aka student info heading) layout in the
feedback list view so that also buttons for viewing conversations with
student are right-justified and near "Background" and "Points" buttons.
This allows all three buttons that open popovers to be near each other
and open the popovers on the right side of the viewport (overlaying feedback
tags instead of conversation) without too much distance to buttons.